[Refactor] DataLayer 리팩토링 - configurator     (As Is)     - DataLayer에서 제공하는 api는 category 관련 api 밖에 없었음.     - DataLayer가 singleton으로, 내부 field로 PageGraphql을 가지고 있었음.. (이렇게 싱글톤으로 만들면 안됐다... 사실 gatsby가 매번 메모리에서 지워서 사실상 싱글톤이 아니었기 때문에 잘 작동 하지 않았나 싶음;;)     - DataLayer 자체가 DI를 수행했다.
    (To Be)
    - DataLayer는 이제 category api 외에도 더 많은 api를 제공할 수 있음
    - DataLayer는 여전히 singleton이지만, 내부 field로 PageGraphql을 지니고 있지 않음.
    - DI 작업은 Configurator에게 위임한다.
      - Configurator가 내부 필드로 PageGraphql을 지닌다.
