+++ src/dataLayer/DataLayer.ts
@@ -1,46 +1,29 @@
 import { PageGraphQL } from 'types/nodeapi-types'
 
 import CategoryAPI from './api/CategoryAPI'
-import MdxPageQueries from './queries/MdxPageQueries'
-import MdxQueries from './queries/mdxQueries'
-import MdxStaticQueries from './queries/MdxStaticQueries'
-import CategoryQueryService from './services/CategoryQueryService'
-import CategoryService from './services/CategoryService'
+import CategoryApiConfigurator from './configurators/CategoryApiConfigurator'
 
 export default class DataLayer {
-  private static DataLayer: DataLayer
-  public readonly API: CategoryAPI
+  private static _singleton: DataLayer
+  private _categoryApi: CategoryAPI
 
-  private constructor(private graphql?: PageGraphQL) {
-    this.API = this._constructCategoryAPI()
+  private constructor() {
+    this._categoryApi = new CategoryApiConfigurator().api
   }
 
-  public static singleton(graphql?: PageGraphQL) {
-    if (!this.DataLayer) {
-      this.DataLayer = new DataLayer(graphql)
+  get categoryApi() {
+    return this._categoryApi
+  }
+
+  static get singleton() {
+    if (!this._singleton) {
+      this._singleton = new DataLayer()
     }
-    return this.DataLayer
+    return this._singleton
   }
 
-  private _constructCategoryAPI() {
-    const mdxQueries = this._getMdxQueries()
-    const categoryQueryService = this._getCategoryQueryService(mdxQueries)
-    const categoryService = this._getCategoryService(categoryQueryService)
-    return new CategoryAPI(categoryService)
-  }
-
-  private _getCategoryService(categoryQueryService: CategoryQueryService) {
-    return new CategoryService(categoryQueryService)
-  }
-
-  private _getCategoryQueryService(mdxQueries: MdxQueries) {
-    return new CategoryQueryService(mdxQueries)
-  }
-
-  private _getMdxQueries() {
-    if (this.graphql) {
-      return new MdxPageQueries(this.graphql)
-    }
-    return new MdxStaticQueries()
+  public configurePageGraphql(graphql: PageGraphQL): DataLayer {
+    this._categoryApi = new CategoryApiConfigurator(graphql).api
+    return this
   }
 }
